Paper dataset for model training
The Excel file contains 4 sheets: Master, 7 promoters, Initial set, Blind Set. Initial set is used as a training set with 14 promoters and 189 datapoints, while the blind set serves as a validation set with 6 metals and 41 datapoints. It should be noted that the paper erroneously states that there are 5 metals and 36 datapoints in the blind set.